Split QUACK epilogue analysis and emission
